radio send binary and ascii mode

+++ DC/src/CAN/CANBusHandlerDC.cpp
@@ -20,7 +20,7 @@ extern CarState carState;
 extern Console console;
 extern CANBus canBus;
 
-bool CANBus::is_to_ignore_packet(int packetId) {
+bool CANBus::is_to_ignore_packet(uint16_t packetId) {
   return packetId != (DC_BASE_ADDR | 0x00) && packetId != (DC_BASE_ADDR | 0x01) && !canBus.isPacketToRenew(packetId);
 }
